requirements.txt
telethon 
pandas 
nltk
python-dotenv
